// part9/ilari-diary/frontend/src/types/types.tsx
export interface Entry {
  id: number;
  date: string;
  weather: string;
  visibility: string;
  comment?: string;
}

export enum Weather {
  Sunny = 'sunny',
  Rainy = 'rainy',
  Cloudy = 'cloudy',
  Stormy = 'stormy',
  Windy = 'windy',
}

export enum Visibility {
  Great = 'great',
  Good = 'good',
  Ok = 'ok',
  Poor = 'poor',
}

export type NewEntry = Omit<Entry, 'id'>;